backend/requirements.txt
# backend/requirements.txt — All packages for all phases

# Django core
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# CORS
django-cors-headers==4.3.1

# Filters + pagination
django-filter==23.5

# Image handling
Pillow==10.2.0

# Environment variables
python-dotenv==1.0.0

# HTTP requests (SSLCommerz calls)
requests==2.31.0

# Static files (production)
whitenoise==6.6.0

# Production WSGI server
gunicorn==21.2.0
